+++ .internal/spindler/spindler/parser.py
@@ -33,10 +33,6 @@ def parse(text) -> list[TweePassage]:
     passages: list[TweePassage] = []
     parsed_twee = twee_grammar.parse(text)
 
-    story_data_node = next(parsed_twee.find_data("story_data"))
-    story_data = json.loads(cast(Token, story_data_node.children[0]).value)
-    start_node = story_data["start"]
-
     for passage_tree in parsed_twee.find_data("passage"):
         name = cast(
             Token, next(passage_tree.find_data("passage_name")).children[0]
@@ -65,7 +61,6 @@ def parse(text) -> list[TweePassage]:
             name=name,
             tree=tree,
             tags=tags,
-            is_start=name == start_node,
         )
         passages.append(passage)